+++ src/main.js
@@ -1,13 +1,20 @@
-import { serialize, parse as parseLib } from 'error-serializer'
-import isErrorInstance from 'is-error-instance'
+import { serialize, parse as parseErrorObject } from 'error-serializer'
 
 const toJSON = function ({ error }) {
   return serialize(error)
 }
 
-const parse = function ({ ErrorClasses, BaseError }, errorObject) {
-  const value = parseLib(errorObject, { classes: ErrorClasses })
-  return isErrorInstance(value) ? BaseError.normalize(value) : value
+const parse = function ({ ErrorClasses }, errorObject) {
+  const classes = getClasses(ErrorClasses)
+  return parseErrorObject(errorObject, { classes })
+}
+
+const getClasses = function (ErrorClasses) {
+  return Object.fromEntries(ErrorClasses.map(getClass))
+}
+
+const getClass = function (ErrorClass) {
+  return [ErrorClass.name, ErrorClass]
 }
 
 export default {