+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - [`modern-errors@5`](https://github.com/ehmicky/modern-errors/releases/tag/5.0.0)
   is now required
+- [`BaseError.parse(value)`](README.md#baseerrorparseerrorobject) is now deep by
+  default. If `value` is not an error object, it is not converted to one
+  anymore.
 
 # 1.4.0